+++ e2e/decision_tree_classifier_test.dart
@@ -69,5 +69,21 @@ void main() async {
 
       expect(scores.mean(), greaterThan(0.5));
     });
+
+    test('should return adequate score on iris dataset using recall '
+        'metric, dtype=DType.float32', () async {
+      final scores = await evaluateClassifier(
+          MetricType.recall, DType.float32);
+
+      expect(scores.mean(), greaterThan(0.5));
+    });
+
+    test('should return adequate score on iris dataset using recall '
+        'metric, dtype=DType.float64', () async {
+      final scores = await evaluateClassifier(
+          MetricType.recall, DType.float64);
+
+      expect(scores.mean(), greaterThan(0.5));
+    });
   });
 }